fastapi-backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
sqlalchemy>=2.0.25
alembic>=1.13.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
anthropic>=0.8.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
pytest>=8.0.0
httpx>=0.26.0
asyncpg>=0.29.0
structlog>=24.1.0
greenlet>=3.1.1

# Development dependencies
black>=24.1.0
isort>=5.13.0
flake8>=7.0.0
mypy>=1.8.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0 